requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
autograd=1.7.0=pypi_0
autograd-gamma=0.5.0=pypi_0
blas=1.0=mkl
bzip2=1.0.8=h7f98852_4
ca-certificates=2024.8.30=hbcca054_0
contourpy=1.3.0=pypi_0
cudatoolkit=11.1.1=ha002fc5_10
cycler=0.12.1=pypi_0
ecos=2.0.14=pypi_0
ffmpeg=4.3=hf484d3e_0
fonttools=4.53.1=pypi_0
formulaic=1.0.2=pypi_0
freetype=2.10.4=h0708190_1
gmp=6.2.1=h58526e2_0
gnutls=3.6.13=h85f3911_1
h5py=3.11.0=pypi_0
importlib-resources=6.4.5=pypi_0
intel-openmp=2021.4.0=h06a4308_3561
interface-meta=1.3.0=pypi_0
joblib=1.4.2=pypi_0
jpeg=9b=h024ee3a_2
kiwisolver=1.4.7=pypi_0
lame=3.100=h7f98852_1001
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.38=h1181459_1
libffi=3.4.4=h6a678d5_1
libgcc-ng=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libiconv=1.17=h166bdaf_0
libpng=1.6.37=h21135ba_2
libstdcxx-ng=11.2.0=h1234567_1
libtiff=4.1.0=h2733197_1
lifelines=0.29.0=pypi_0
lz4-c=1.9.3=h9c3ff4c_1
matplotlib=3.9.2=pypi_0
mkl=2021.4.0=h06a4308_640
mkl-service=2.4.0=py39h7e14d7c_0
mkl_fft=1.3.1=py39h0c7bc48_1
mkl_random=1.2.2=py39hde0f152_0
ncurses=6.4=h6a678d5_0
nettle=3.6=he412f7d_0
ninja=1.11.0=h924138e_0
numexpr=2.10.1=pypi_0
numpy=1.26.4=pypi_0
numpy-base=1.24.3=py39h31eccc5_0
olefile=0.47=pyhd8ed1ab_0
openh264=2.1.1=h780b84a_0
openssl=3.0.15=h5eee18b_0
osqp=0.6.7.post1=pypi_0
packaging=24.1=pypi_0
pandas=2.2.2=pypi_0
pillow=10.4.0=pypi_0
pip=24.2=py39h06a4308_0
protobuf=5.28.1=pypi_0
pyparsing=3.1.4=pypi_0
python=3.9.16=h955ad1f_3
python-dateutil=2.9.0.post0=pypi_0
python_abi=3.9=2_cp39
pytorch=1.8.0=py3.9_cuda11.1_cudnn8.0.5_0
pytz=2024.2=pypi_0
qdldl=0.1.7.post4=pypi_0
readline=8.2=h5eee18b_0
scikit-learn=1.5.2=pypi_0
scikit-survival=0.23.0=pypi_0
scipy=1.13.1=pypi_0
setuptools=72.1.0=py39h06a4308_0
six=1.16.0=pyh6c4a22f_0
sqlite=3.45.3=h5eee18b_0
tensorboardx=2.6.2.2=pypi_0
threadpoolctl=3.5.0=pypi_0
tk=8.6.14=h39e8969_0
torchvision=0.9.0=py39_cu111
typing_extensions=4.12.2=pyha770c72_0
tzdata=2024.1=pypi_0
wheel=0.44.0=py39h06a4308_0
wrapt=1.16.0=pypi_0
xz=5.4.6=h5eee18b_1
zipp=3.20.1=pypi_0
zlib=1.2.13=h5eee18b_1
zstd=1.4.9=ha95c52a_0
